move tree store
