+++ src/packages/settings/data-types/tree/data-type.tree.store.ts
@@ -6,9 +6,8 @@ import { UmbEntityTreeStore } from '@umbraco-cms/backoffice/store';
  * @export
  * @class UmbDataTypeTreeStore
  * @extends {UmbStoreBase}
- * @description - Tree Data Store for Data-Types
+ * @description - Tree Data Store for Data Types
  */
-// TODO: consider if tree store could be turned into a general EntityTreeStore class?
 export class UmbDataTypeTreeStore extends UmbEntityTreeStore {
 	/**
 	 * Creates an instance of UmbDataTypeTreeStore.
